[QoS] Guava license (#2703)
* guava license

* Cleanup: class reference
