--- qos-service-impl/src/main/java/com/palantir/atlasdb/qos/ratelimit/RateLimiter.java
@@ -1,306 +1,0 @@
-//CHECKSTYLE:OFF
-/*
- * Copyright (C) 2012 The Guava Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.atlasdb.qos.ratelimit;
-
-import static java.lang.Math.max;
-import static java.util.concurrent.TimeUnit.MICROSECONDS;
-import static java.util.concurrent.TimeUnit.SECONDS;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkNotNull;
-
-import java.time.Duration;
-import java.util.Locale;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-
-import com.google.common.base.Stopwatch;
-import com.google.common.util.concurrent.Uninterruptibles;
-
-/**
- * Copied from Guava, because {@link com.palantir.atlasdb.qos.ratelimit.SmoothRateLimiter.SmoothBursty} is a package
- * private class.
- *
- * There are also a few minor but notable modifications:
- * 1) {@link #tryAcquire(int, long, TimeUnit)} returns an optional duration rather than a boolean. This is analogous
- *    to the return value of {@link #acquire()}.
- * 2) A new method {@link #steal(int)} was added, to support taking permits without waiting
- * 3) Some static constructors were removed.
- **/
-public abstract class RateLimiter {
-
-    /**
-     * The underlying timer; used both to measure elapsed time and sleep as necessary. A separate
-     * object to facilitate testing.
-     */
-    private final SleepingStopwatch stopwatch;
-
-    // Can't be initialized in the constructor because mocks don't call the constructor.
-    private volatile Object mutexDoNotUseDirectly;
-
-    private Object mutex() {
-        Object mutex = mutexDoNotUseDirectly;
-        if (mutex == null) {
-            synchronized (this) {
-                mutex = mutexDoNotUseDirectly;
-                if (mutex == null) {
-                    mutexDoNotUseDirectly = mutex = new Object();
-                }
-            }
-        }
-        return mutex;
-    }
-
-    RateLimiter(SleepingStopwatch stopwatch) {
-        this.stopwatch = checkNotNull(stopwatch);
-    }
-
-    /**
-     * Updates the stable rate of this {@code RateLimiter}, that is, the {@code permitsPerSecond}
-     * argument provided in the factory method that constructed the {@code RateLimiter}. Currently
-     * throttled threads will <b>not</b> be awakened as a result of this invocation, thus they do not
-     * observe the new rate; only subsequent requests will.
-     *
-     * <p>Note though that, since each request repays (by waiting, if necessary) the cost of the
-     * <i>previous</i> request, this means that the very next request after an invocation to
-     * {@code setRate} will not be affected by the new rate; it will pay the cost of the previous
-     * request, which is in terms of the previous rate.
-     *
-     * <p>The behavior of the {@code RateLimiter} is not modified in any other way, e.g. if the
-     * {@code RateLimiter} was configured with a warmup period of 20 seconds, it still has a warmup
-     * period of 20 seconds after this method invocation.
-     *
-     * @param permitsPerSecond the new stable rate of this {@code RateLimiter}
-     * @throws IllegalArgumentException if {@code permitsPerSecond} is negative or zero
-     */
-    public final void setRate(double permitsPerSecond) {
-        checkArgument(
-                permitsPerSecond > 0.0 && !Double.isNaN(permitsPerSecond), "rate must be positive");
-        synchronized (mutex()) {
-            doSetRate(permitsPerSecond, stopwatch.readMicros());
-        }
-    }
-
-    abstract void doSetRate(double permitsPerSecond, long nowMicros);
-
-    /**
-     * Returns the stable rate (as {@code permits per seconds}) with which this {@code RateLimiter} is
-     * configured with. The initial value of this is the same as the {@code permitsPerSecond} argument
-     * passed in the factory method that produced this {@code RateLimiter}, and it is only updated
-     * after invocations to {@linkplain #setRate}.
-     */
-    public final double getRate() {
-        synchronized (mutex()) {
-            return doGetRate();
-        }
-    }
-
-    abstract double doGetRate();
-
-    /**
-     * Acquires a single permit from this {@code RateLimiter}, blocking until the request can be
-     * granted. Tells the amount of time slept, if any.
-     *
-     * <p>This method is equivalent to {@code acquire(1)}.
-     *
-     * @return time spent sleeping to enforce rate, in seconds; 0.0 if not rate-limited
-     * @since 16.0 (present in 13.0 with {@code void} return type})
-     */
-    public double acquire() {
-        return acquire(1);
-    }
-
-    /**
-     * Acquires the given number of permits from this {@code RateLimiter}, blocking until the request
-     * can be granted. Tells the amount of time slept, if any.
-     *
-     * @param permits the number of permits to acquire
-     * @return time spent sleeping to enforce rate, in seconds; 0.0 if not rate-limited
-     * @throws IllegalArgumentException if the requested number of permits is negative or zero
-     * @since 16.0 (present in 13.0 with {@code void} return type})
-     */
-    public double acquire(int permits) {
-        long microsToWait = reserve(permits);
-        stopwatch.sleepMicrosUninterruptibly(microsToWait);
-        return 1.0 * microsToWait / SECONDS.toMicros(1L);
-    }
-
-    /**
-     * Immediately steals the given number of permits. This will potentially penalize future callers, but has no
-     * effect on callers that are already waiting for permits.
-     */
-    public void steal(int permits) {
-        reserve(permits);
-    }
-
-    /**
-     * Reserves the given number of permits from this {@code RateLimiter} for future use, returning
-     * the number of microseconds until the reservation can be consumed.
-     *
-     * @return time in microseconds to wait until the resource can be acquired, never negative
-     */
-    final long reserve(int permits) {
-        checkPermits(permits);
-        synchronized (mutex()) {
-            return reserveAndGetWaitLength(permits, stopwatch.readMicros());
-        }
-    }
-
-    /**
-     * Acquires a permit from this {@code RateLimiter} if it can be obtained without exceeding the
-     * specified {@code timeout}, or returns {@code false} immediately (without waiting) if the permit
-     * would not have been granted before the timeout expired.
-     *
-     * <p>This method is equivalent to {@code tryAcquire(1, timeout, unit)}.
-     *
-     * @param timeout the maximum time to wait for the permit. Negative values are treated as zero.
-     * @param unit the time unit of the timeout argument
-     * @return {@code true} if the permit was acquired, {@code false} otherwise
-     * @throws IllegalArgumentException if the requested number of permits is negative or zero
-     */
-    public boolean tryAcquire(long timeout, TimeUnit unit) {
-        return tryAcquire(1, timeout, unit).isPresent();
-    }
-
-    /**
-     * Acquires permits from this {@link com.google.common.util.concurrent.RateLimiter} if it can be acquired immediately without delay.
-     *
-     * <p>This method is equivalent to {@code tryAcquire(permits, 0, anyUnit)}.
-     *
-     * @param permits the number of permits to acquire
-     * @return {@code true} if the permits were acquired, {@code false} otherwise
-     * @throws IllegalArgumentException if the requested number of permits is negative or zero
-     * @since 14.0
-     */
-    public boolean tryAcquire(int permits) {
-        return tryAcquire(permits, 0, MICROSECONDS).isPresent();
-    }
-
-    /**
-     * Acquires a permit from this {@link com.google.common.util.concurrent.RateLimiter} if it can be acquired immediately without
-     * delay.
-     *
-     * <p>This method is equivalent to {@code tryAcquire(1)}.
-     *
-     * @return {@code true} if the permit was acquired, {@code false} otherwise
-     * @since 14.0
-     */
-    public boolean tryAcquire() {
-        return tryAcquire(1, 0, MICROSECONDS).isPresent();
-    }
-
-    /**
-     * Acquires the given number of permits from this {@code RateLimiter} if it can be obtained
-     * without exceeding the specified {@code timeout}, or returns {@code false} immediately (without
-     * waiting) if the permits would not have been granted before the timeout expired.
-     *
-     * @param permits the number of permits to acquire
-     * @param timeout the maximum time to wait for the permits. Negative values are treated as zero.
-     * @param unit the time unit of the timeout argument
-     * @return amount of time waited, if the permits were acquired, empty otherwise
-     * @throws IllegalArgumentException if the requested number of permits is negative or zero
-     */
-    public Optional<Duration> tryAcquire(int permits, long timeout, TimeUnit unit) {
-        long timeoutMicros = max(unit.toMicros(timeout), 0);
-        checkPermits(permits);
-        long microsToWait;
-        synchronized (mutex()) {
-            long nowMicros = stopwatch.readMicros();
-            if (!canAcquire(nowMicros, timeoutMicros)) {
-                return Optional.empty();
-            } else {
-                microsToWait = reserveAndGetWaitLength(permits, nowMicros);
-            }
-        }
-        stopwatch.sleepMicrosUninterruptibly(microsToWait);
-        return Optional.of(Duration.ofNanos(TimeUnit.MICROSECONDS.toNanos(microsToWait)));
-    }
-
-    private boolean canAcquire(long nowMicros, long timeoutMicros) {
-        return queryEarliestAvailable(nowMicros) - timeoutMicros <= nowMicros;
-    }
-
-    /**
-     * Reserves next ticket and returns the wait time that the caller must wait for.
-     *
-     * @return the required wait time, never negative
-     */
-    final long reserveAndGetWaitLength(int permits, long nowMicros) {
-        long momentAvailable = reserveEarliestAvailable(permits, nowMicros);
-        return max(momentAvailable - nowMicros, 0);
-    }
-
-    /**
-     * Returns the earliest time that permits are available (with one caveat).
-     *
-     * @return the time that permits are available, or, if permits are available immediately, an
-     *     arbitrary past or present time
-     */
-    abstract long queryEarliestAvailable(long nowMicros);
-
-    /**
-     * Reserves the requested number of permits and returns the time that those permits can be used
-     * (with one caveat).
-     *
-     * @return the time that the permits may be used, or, if the permits may be used immediately, an
-     *     arbitrary past or present time
-     */
-    abstract long reserveEarliestAvailable(int permits, long nowMicros);
-
-    @Override
-    public String toString() {
-        return String.format(Locale.ROOT, "RateLimiter[stableRate=%3.1fqps]", getRate());
-    }
-
-    abstract static class SleepingStopwatch {
-        /** Constructor for use by subclasses. */
-        protected SleepingStopwatch() {}
-
-        /*
-         * We always hold the mutex when calling this. TODO(cpovirk): Is that important? Perhaps we need
-         * to guarantee that each call to reserveEarliestAvailable, etc. sees a value >= the previous?
-         * Also, is it OK that we don't hold the mutex when sleeping?
-         */
-        protected abstract long readMicros();
-
-        protected abstract void sleepMicrosUninterruptibly(long micros);
-
-        public static final SleepingStopwatch createFromSystemTimer() {
-            return new SleepingStopwatch() {
-                final Stopwatch stopwatch = Stopwatch.createStarted();
-
-                @Override
-                protected long readMicros() {
-                    return stopwatch.elapsed(MICROSECONDS);
-                }
-
-                @Override
-                protected void sleepMicrosUninterruptibly(long micros) {
-                    if (micros > 0) {
-                        Uninterruptibles.sleepUninterruptibly(micros, MICROSECONDS);
-                    }
-                }
-            };
-        }
-    }
-
-    private static void checkPermits(int permits) {
-        checkArgument(permits > 0, "Requested permits (%s) must be positive", permits);
-    }
-}
-
