--- qos-service-impl/src/main/java/com/palantir/atlasdb/qos/ratelimit/SmoothRateLimiter.java
@@ -1,154 +1,0 @@
-// CHECKSTYLE:OFF
-/*
- * Copyright (C) 2012 The Guava Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.atlasdb.qos.ratelimit;
-
-import static java.lang.Math.min;
-import static java.util.concurrent.TimeUnit.SECONDS;
-
-/**
- * Copied from Guava, because {@link SmoothBursty} is a package private class.
- **/
-abstract class SmoothRateLimiter extends RateLimiter {
-    /**
-     * This implements a "bursty" RateLimiter, where storedPermits are translated to zero throttling.
-     * The maximum number of permits that can be saved (when the RateLimiter is unused) is defined in
-     * terms of time, in this sense: if a RateLimiter is 2qps, and this time is specified as 10
-     * seconds, we can save up to 2 * 10 = 20 permits.
-     */
-    static final class SmoothBursty extends SmoothRateLimiter {
-        /** The work (permits) of how many seconds can be saved up if this RateLimiter is unused? */
-        final double maxBurstSeconds;
-
-        SmoothBursty(SleepingStopwatch stopwatch, double maxBurstSeconds) {
-            super(stopwatch);
-            this.maxBurstSeconds = maxBurstSeconds;
-        }
-
-        @Override
-        void doSetRate(double permitsPerSecond, double stableIntervalMicros) {
-            double oldMaxPermits = this.maxPermits;
-            maxPermits = maxBurstSeconds * permitsPerSecond;
-            if (oldMaxPermits == Double.POSITIVE_INFINITY) {
-                // if we don't special-case this, we would get storedPermits == NaN, below
-                storedPermits = maxPermits;
-            } else {
-                storedPermits =
-                        (oldMaxPermits == 0.0)
-                                ? 0.0 // initial state
-                                : storedPermits * maxPermits / oldMaxPermits;
-            }
-        }
-
-        @Override
-        long storedPermitsToWaitTime(double storedPermits, double permitsToTake) {
-            return 0L;
-        }
-
-        @Override
-        double coolDownIntervalMicros() {
-            return stableIntervalMicros;
-        }
-    }
-
-    /**
-     * The currently stored permits.
-     */
-    double storedPermits;
-
-    /**
-     * The maximum number of stored permits.
-     */
-    double maxPermits;
-
-    /**
-     * The interval between two unit requests, at our stable rate. E.g., a stable rate of 5 permits
-     * per second has a stable interval of 200ms.
-     */
-    double stableIntervalMicros;
-
-    /**
-     * The time when the next request (no matter its size) will be granted. After granting a request,
-     * this is pushed further in the future. Large requests push this further than small requests.
-     */
-    private long nextFreeTicketMicros = 0L; // could be either in the past or future
-
-    private SmoothRateLimiter(SleepingStopwatch stopwatch) {
-        super(stopwatch);
-    }
-
-    @Override
-    final void doSetRate(double permitsPerSecond, long nowMicros) {
-        resync(nowMicros);
-        double stableIntervalMicros = SECONDS.toMicros(1L) / permitsPerSecond;
-        this.stableIntervalMicros = stableIntervalMicros;
-        doSetRate(permitsPerSecond, stableIntervalMicros);
-    }
-
-    abstract void doSetRate(double permitsPerSecond, double stableIntervalMicros);
-
-    @Override
-    final double doGetRate() {
-        return SECONDS.toMicros(1L) / stableIntervalMicros;
-    }
-
-    @Override
-    final long queryEarliestAvailable(long nowMicros) {
-        return nextFreeTicketMicros;
-    }
-
-    @Override
-    final long reserveEarliestAvailable(int requiredPermits, long nowMicros) {
-        resync(nowMicros);
-        long returnValue = nextFreeTicketMicros;
-        double storedPermitsToSpend = min(requiredPermits, this.storedPermits);
-        double freshPermits = requiredPermits - storedPermitsToSpend;
-        long waitMicros =
-                storedPermitsToWaitTime(this.storedPermits, storedPermitsToSpend)
-                        + (long) (freshPermits * stableIntervalMicros);
-
-        this.nextFreeTicketMicros = nextFreeTicketMicros + waitMicros;
-        this.storedPermits -= storedPermitsToSpend;
-        return returnValue;
-    }
-
-    /**
-     * Translates a specified portion of our currently stored permits which we want to spend/acquire,
-     * into a throttling time. Conceptually, this evaluates the integral of the underlying function we
-     * use, for the range of [(storedPermits - permitsToTake), storedPermits].
-     *
-     * <p>This always holds: {@code 0 <= permitsToTake <= storedPermits}
-     */
-    abstract long storedPermitsToWaitTime(double storedPermits, double permitsToTake);
-
-    /**
-     * Returns the number of microseconds during cool down that we have to wait to get a new permit.
-     */
-    abstract double coolDownIntervalMicros();
-
-    /**
-     * Updates {@code storedPermits} and {@code nextFreeTicketMicros} based on the current time.
-     */
-    void resync(long nowMicros) {
-        // if nextFreeTicket is in the past, resync to now
-        if (nowMicros > nextFreeTicketMicros) {
-            double newPermits = (nowMicros - nextFreeTicketMicros) / coolDownIntervalMicros();
-            storedPermits = min(maxPermits, storedPermits + newPermits);
-            nextFreeTicketMicros = nowMicros;
-        }
-    }
-}
